test: refactor
refs #8
