+++ src/pages/about-me/index.tsx
@@ -6,7 +6,7 @@ import type { HeadFC } from 'gatsby';
 import { ExternalLink } from '../../components/external-link/external-link';
 import { InternalLink } from '../../components/internal-link/internal-link';
 import { Layout } from '../../components/layout/layout';
-import { Seo } from '../../components/seo';
+import { Seo } from '../../components/seo/seo';
 
 const AboutMePage: FunctionComponent = (): ReactElement => {
     return (
